requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
jsons==1.6.3
numpy==1.26.4
opencv-python==4.9.0.80
PyYAML==6.0.1
requests==2.31.0
typish==1.9.3
urllib3==2.2.1
datetime